+++ src/components/study-card/study-categories.js
@@ -1,0 +1,29 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+export default {
+    "social media": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#F5F1BA"
+    },
+    "misinformation": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#FFD5B2"
+    },
+    "artificial intelligence": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#D98FFF"
+    },
+    "devices": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#FFB4DC"
+    },
+    "data & privacy": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#B6F4F7"
+    },
+    "accessibility": {
+        text: "var(--color-marketing-gray-100)",
+        background: "#E0ECEC"
+    },
+}